requirements.txt
asgiref==3.7.2
Django==4.2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.5.0
sqlparse==0.4.4
typing_extensions==4.6.3
